+++ lib/src/Controller/currency_controller.dart
@@ -11,6 +11,7 @@ class CurrencyController{
 
   ValueNotifier<double?> mount = ValueNotifier<double?>(0);
   ValueNotifier<Currency?> _currency = ValueNotifier<Currency?>(null);
+  ValueNotifier<Currency?> get currencyNotifier => _currency;
   Currency get currency => _currency.value??supportedCurrencies[0];
   void set currency(Currency? currency){
     _currency.value = currency;